Add more commands to commands/
